Separate `includeSources` to another file
